requirements.txt
Pillow==6.0.0
mutagen==1.42.0
PyPDF2==1.26.0
filetype==1.0.5